==> 01_itunes_scrape_popular_podcasts.ipynb <==
# Popular podcasts on iTunes
In this notebook, I collect the iTunes id of all the podcasts that are listed in the popular podcasts page on the iTunes website. There are 19 main categories and 102 sub-categories. In total, I find about 22000 unique podcasts and save them in a pandas dataframe.

import pandas as pd
import requests
import json
from bs4 import BeautifulSoup
import numpy as np
import subprocess
from dateutil.parser import parse
import datetime
import re

# get the url for each popular category
re = requests.get(url = 'https://podcasts.apple.com/us/genre/podcasts-arts/id1301')
soup = BeautifulSoup(re.text)
classes = soup.find_all('ul', attrs= {"class":"list column first"})[0].\
    find_all('a', attrs = {"class":"top-level-genre"})
genre_links = {pod_class.text:pod_class['href'] for pod_class in classes}

genre_links

all_links = dict()
for genre, genre_link in genre_links.items():
    re = requests.get(genre_link)
    soup = BeautifulSoup(re.text)
    try:
        sub_genres = soup.find_all('ul', attrs= {"class":"top-level-subgenres"})[0].find_all('a')
        subgenre_links = {subs.text:subs['href'] for subs in sub_genres}
    except:
        subgenre_links = {genre:genre_link}
    all_links[genre] = subgenre_links

itunes_ids = []
genres = []
subgenres = []

for genre in all_links.keys():
    print(genre, end = ', ')
    url = genre_links[genre]
    re = requests.get(url)
    soup = BeautifulSoup(re.text)
    pods = soup.find_all('div', attrs= {"class":"grid3-column"})[0].find_all("a")
    for pod in pods:
        s = pod['href'][::-1]
        idx = s.find("di")
        itunes_ids.append(int(s[:idx][::-1]))
        genres.append(genre)   
        subgenres.append(genre)

    for subgenre, url in all_links[genre].items():
        re = requests.get(url)
        soup = BeautifulSoup(re.text)
        pods = soup.find_all('div', attrs= {"class":"grid3-column"})[0].find_all("a")
        for pod in pods:
            s = pod['href'][::-1]
            idx = s.find("di")
            itunes_ids.append(int(s[:idx][::-1]))
            genres.append(genre)   
            subgenres.append(subgenre)

df = pd.DataFrame(columns = ["itunes_id", "genre", "subgenre"])
df["itunes_id"] = itunes_ids
df["genre"] = genres
df["subgenre"] = subgenres

popular_pods_list = df.loc[df["itunes_id"].drop_duplicates(keep = "last").index].reset_index(drop = True)
popular_pods_list

popular_pods_list.to_pickle("popular_pods_list.pkl")
