==> 02_itunes_scrape_episodes.ipynb <==
# Scraping episodes data from iTunes

In this notebook, I collect episodes data for each podcast that was found in the previous notebook. For each podcast, all the episodes that are available on iTunes website are found (up to 300 episodes for each podcast). For each episode, I collect these attributes:
 - Name
 - Description
 - Release date
 - Duration
 
For each podcast, iTunes proved a list of podcasts that the listeners also subscribed too (up to 18 items). I collect these as well and add them to the popular podcasts dataframe.

import pandas as pd
import requests
import json
from bs4 import BeautifulSoup
import numpy as np
import subprocess
from dateutil.parser import parse
import datetime
import re

alsoSubbedCmd_0 = """curl 'https://amp-api.podcasts.apple.com/v1/catalog/us/podcasts/{}/listeners-also-subscribed' \
  -H 'Accept: application/json' \
  -H 'Referer: https://podcasts.apple.com/podcast/id{}' \
  -H 'Authorization: Bearer eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IldlYlBsYXlLaWQifQ.eyJpc3MiOiJBTVBXZWJQbGF5IiwiaWF0IjoxNTk4NjQzODE5LCJleHAiOjE2MTQxOTU4MTl9.D83tGI4HTOgGUgvKbnyjYtTGxqb7HcrUFixvcMHXvQmSqNy7TOgxO1WIqeBpnql7ibqMdUEImKbNgzfC671qjg' \
  -H 'User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36' \
  -H 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8' \
  --compressed"""

alsoSubbedCmd_1 = """curl 'https://amp-api.podcasts.apple.com/v1/catalog/us/podcasts/{}/listeners-also-subscribed?offset=15' \
  -H 'Accept: application/json' \
  -H 'Referer: https://podcasts.apple.com/podcast/id{}' \
  -H 'Authorization: Bearer eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IldlYlBsYXlLaWQifQ.eyJpc3MiOiJBTVBXZWJQbGF5IiwiaWF0IjoxNTk4NjQzODE5LCJleHAiOjE2MTQxOTU4MTl9.D83tGI4HTOgGUgvKbnyjYtTGxqb7HcrUFixvcMHXvQmSqNy7TOgxO1WIqeBpnql7ibqMdUEImKbNgzfC671qjg' \
  -H 'User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36' \
  -H 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8' \
  --compressed"""

def get_also_subbed_data(_id):
    out1 = subprocess.Popen(alsoSubbedCmd_0.format(_id, _id), stdout=subprocess.PIPE, shell = True).communicate()[0]
    out2 = subprocess.Popen(alsoSubbedCmd_1.format(_id, _id), stdout=subprocess.PIPE, shell = True).communicate()[0]
    try:
        data1 = json.loads(out1)['data']
        data2 = json.loads(out2)['data']
        data1.extend(data2)
        return [d['id'] for d in data1]
    except:
        return None

get_also_subbed_data(1530405817)

episodes_data = """curl 'https://amp-api.podcasts.apple.com/v1/catalog/us/podcasts/{}/episodes?offset={}&limit={}' \
  -H 'authority: amp-api.podcasts.apple.com' \
  -H 'accept: application/json' \
  -H 'authorization: Bearer eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6IldlYlBsYXlLaWQifQ.eyJpc3MiOiJBTVBXZWJQbGF5IiwiaWF0IjoxNTk4NjQzODE5LCJleHAiOjE2MTQxOTU4MTl9.D83tGI4HTOgGUgvKbnyjYtTGxqb7HcrUFixvcMHXvQmSqNy7TOgxO1WIqeBpnql7ibqMdUEImKbNgzfC671qjg' \
  -H 'user-agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36' \
  -H 'content-type: application/x-www-form-urlencoded; charset=UTF-8' \
  -H 'origin: https://podcasts.apple.com' \
  -H 'sec-fetch-site: same-site' \
  -H 'sec-fetch-mode: cors' \
  -H 'sec-fetch-dest: empty' \
  -H 'referer: https://podcasts.apple.com/podcast/id{}' \
  -H 'accept-language: en-US,en;q=0.9,fa;q=0.8' \
  --compressed"""

def get_episode_data(_id, offset, lim):
    out = subprocess.Popen(episodes_data.format(_id, offset, lim, _id), stdout=subprocess.PIPE, shell = True).communicate()[0]
    data = json.loads(out)
    if 'next' in data:
        idx = data['next'].find("=")
        next_offset = data['next'][idx+1:]
    else:
        next_offset = None
    descriptions = []
    titles = []
    durations = []
    dates = []
    if 'data' in data:
        for d in data['data']:
            try:
                descriptions.append(d['attributes']['description']['standard'])
            except:
                descriptions.append(None)
            try:
                titles.append(d['attributes']['name'])
            except:
                titles.append(None)
            try:
                durations.append(d['attributes']['durationInMilliseconds'])
            except:
                durations.append(None)
            try:
                dates.append(parse(d['attributes']['releaseDateTime']))
            except:
                dates.append(None)
    return titles, descriptions, durations, dates, next_offset


def get_all_episodes(_id):
    k = 0
    descriptions = []
    durations = []
    dates = []
    titles = []
    while k is not None:
        t, de, du, da, k = get_episode_data(_id, k, 100)
        titles.extend(t)        
        descriptions.extend(de)
        durations.extend(du)
        dates.extend(da)
    return titles, descriptions, durations, dates

titles, descriptions, durations, dates = get_all_episodes(1200361736)

pod_ids = pd.read_pickle("popular_pods_list.pkl")

pod_groups = pod_ids.groupby(by = "subgenre").groups

subgenres = list(pod_groups.keys())

for current in subgenres:
    print("_".join(current.lower().split()))
    df = pd.DataFrame(columns = ["itunes_id", "title", "description", "duration", "date"])
    current_pods = pod_ids.loc[pod_groups[current]]
    for count, idx in enumerate(current_pods.index):
        if (count+1)%10 == 0:
            print("{}/{}".format(count+1, current_pods.shape[0]), end = ' | ')
        _id = current_pods.loc[idx, "itunes_id"]
        titles, descriptions, durations, dates = get_all_episodes(_id)
        df = df.append(pd.DataFrame({"itunes_id":_id, "title":titles, "description":descriptions,
                      "duration":durations, "date":dates}))
    file_name = "episodes_subgenre_" + "_".join(current.lower().split()) + '.pkl'
    df.reset_index(drop = True, inplace=True)
    df.to_pickle(file_name)
    print()

also_subbed = pod_ids['itunes_id'].apply(get_also_subbed_data)

pod_ids["also_subbed"] = also_subbed

pod_ids.to_pickle("popular_pods_list_with_also_subbed.pkl")

pod_ids
