==> 03_episodes_word2vec.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
import re
import spacy
from spacy.lang.en.stop_words import STOP_WORDS
nlp = spacy.load("en_core_web_lg", disable=['parser','ner'])

def get_vectors(text):
#     remove mixed alphanumeric
    text = re.sub(r"(?x)\b(?=\w*\d)\w+\s*","", text)
#     remove urls
    text = re.sub(r'(?P<url>https?://[^\s]+)', '', text)
#     remove non-alphanumeric, non-space
    text = re.sub(r'([^\s\w]|_)+', '', text)
    return np.array(nlp(text).vector)

pod_ids = pd.read_pickle("popular_pods_list.pkl")
pod_groups = pod_ids.groupby(by = "subgenre").groups
subgenres = list(pod_groups.keys())

for current in subgenres:
    temp = "_".join(current.lower().split())
    print(temp)
    file_name = "episodes_subgenre_" + temp + '.pkl'
    df = pd.read_pickle(file_name)
    df.dropna(subset=["description"], inplace=True)
    df.reset_index(drop = True, inplace = True)
    chunk_size = 500
    for idx in range(0, df.shape[0], chunk_size):
        df.loc[idx:idx+chunk_size, "vector"] = df["description"][idx:idx+chunk_size].apply(get_vectors)
        print("{}/{}".format(idx+chunk_size, df.shape[0]), end = ' | ')    
    print()
    file_name = "episodes_vector_subgenre_" + temp + '.pkl'
    df.to_pickle(file_name)
