==> 04_filter_lang_get_mean_vector.ipynb <==
import pandas as pd
import numpy as np
from langdetect import detect

pod_ids = pd.read_pickle("popular_pods_list.pkl")
pod_groups = pod_ids.groupby(by = "subgenre").groups
subgenres = list(pod_groups.keys())

def lang_detect(x):
    try:
        return detect(" ".join(x.split()))
    except:
        return None

pods_mean_vec = pd.DataFrame(columns = ["itunes_id", "vector"])
for current in subgenres:
    temp = "_".join(current.lower().split())
    print(temp)
    file_name = "episode_vectors/episodes_vector_subgenre_" + temp + '.pkl'
    df = pd.read_pickle(file_name)
    chunk_size = 1000
    for idx in range(0, df.shape[0], chunk_size):
        df.loc[idx:idx+chunk_size, "language"] = df["description"][idx:idx+chunk_size].apply(lang_detect)
        print("{}/{}".format(idx+chunk_size, df.shape[0]), end = ' | ')    
    df.dropna(subset=["language"], inplace = True)
    df = df.groupby(by = "itunes_id").filter(lambda x:len(x)>5).reset_index(drop = True)
    mean_vec = df[["itunes_id", "vector"]].groupby(by = "itunes_id").apply(np.mean)
    mean_vec['itunes_id'] = mean_vec.index
    mean_vec.reset_index(drop = True)
    pods_mean_vec = pods_mean_vec.append(mean_vec)
    print(mean_vec.shape[0])    

pods_mean_vec

pods_mean_vec.to_pickle("pods_mean_vec.pkl")
