==> 05_scrape_podcasts_descriptions.ipynb <==
import pandas as pd
import requests
import json
from bs4 import BeautifulSoup
import numpy as np
import subprocess
from dateutil.parser import parse
import datetime
import re

df = pd.read_pickle("popular_pods_list.pkl")

new_ids = df["itunes_id"].to_list()

%%time
# get podcast descriptions, titles, ratings, and list of related ones
descriptions = []
titles = []
for i, itunes_id in enumerate(new_ids):
    if (i+1)%10 == 0:
        print("{}/{}".format(i+1, len(new_ids)), end = ' | ')
    url = "https://podcasts.apple.com/podcast/id" + str(itunes_id)
    re = requests.get(url)
    soup = BeautifulSoup(re.text)
    try:
        titles.append(" ".join(soup.find("span", attrs = {"class":"product-header__title"}).text.split()))
    except:
        titles.append(None)
    try:
        descriptions.append(soup.find("section", attrs = {"class":"product-hero-desc__section"}).find('p').text)
    except:
        descriptions.append(None)

df["titles"] = titles
df["descriptions"] = descriptions
df

df.to_pickle("pod_descriptions.pkl")

df["descriptions"].isna().value_counts()
