==> 07_itunes_scrape_artwork_url.ipynb <==
import pandas as pd
import numpy as np
import requests
import json

pods = pd.read_pickle("pod_descriptions.pkl")

pods = pods[["itunes_id"]]

%%time
artwork_urls = []
for i, _id in enumerate(pods["itunes_id"]):
    if (i+1)%100 == 0:
        print(i+1, end = " | ")
    try:
        req = requests.get("https://itunes.apple.com/lookup?id=" + str(_id))
        res = json.loads(req.text)
        artwork_urls.append(res['results'][0]['artworkUrl600'])
    except:
        artwork_urls.append(None)

pods = pd.read_pickle("pod_descriptions.pkl")

pods['artwork_url'] = artwork_urls

pods

pods.to_pickle("pod_descriptions.pkl")
