==> 08_prepare_tables.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity

df = pd.read_pickle("similar_pods_based_on_episodes.pkl")

pods1 = pd.read_pickle("pod_descriptions.pkl")
pods2 = pd.read_pickle("pods_mean_vec_description.pkl")

temp = pods2[["itunes_id","vector"]].merge(pods1, on = "itunes_id").merge(df, on = "itunes_id")[["itunes_id", "matches"]]

_ids = []
match_id = []
score = []
for i in range(temp.shape[0]):
    if (i+1)%100 == 0:
        print(i+1, end = " | ")
    current_matches = temp.loc[i,"matches"]
    match_id.extend(current_matches["itunes_id"].to_list())
    score.extend(current_matches["score"].to_list())
    _ids.extend([temp.loc[i,"itunes_id"]]*(current_matches.shape[0]))

all_matches = pd.DataFrame(columns=["itunes_id","match_id","score"])
all_matches["itunes_id"] = _ids
all_matches["match_id"] = match_id
all_matches["score"] = score

all_pods = pods2[["itunes_id","vector"]].merge(pods1, on = "itunes_id").merge(df, on = "itunes_id")\
    [["itunes_id", "titles", "descriptions", "genre", "subgenre", "artwork_url"]]

vecs = pods2[["itunes_id","vector"]].merge(pods1, on = "itunes_id").merge(df, on = "itunes_id")["vector"].to_numpy()
vecs = np.array([vec for vec in vecs])

sims = cosine_similarity(vecs)
top100 = np.argpartition(-sims, 101)[:,:101]
val = np.array([sims[i,:][top100[i,:]] for i in range(sims.shape[0])])
top100_idx = np.argsort(-val)
matches_meanvec = np.array([top100[i,top100_idx[i]] for i in range(sims.shape[0])])
scores_meanvec = np.array([val[i,top100_idx[i]] for i in range(sims.shape[0])])
matches_meanvec = matches_meanvec[:,1:]
scores_meanvec = scores_meanvec[:,1:]
matches_meanvec = matches_meanvec.reshape(19710*100)
scores_meanvec = scores_meanvec.reshape(19710*100)

ids = all_pods["itunes_id"].to_list()
id_list = [_id for _id in ids for i in range(100)]

all_matches_meanvec = pd.DataFrame(columns=["itunes_id","match_id","score"])
all_matches_meanvec["itunes_id"] = id_list
all_matches_meanvec["match_id"] = all_pods["itunes_id"].loc[matches_meanvec].to_list()
all_matches_meanvec["score"] = scores_meanvec

all_matches.to_pickle("all_matches_episode_based.pkl")
all_matches_meanvec.to_pickle("all_matches_mean_vec_based.pkl")
all_pods.to_pickle("all_pods.pkl")

all_pods[all_pods["titles"] == "The Liberty Yell"]

idx = 16764 #np.random.randint(all_pods.shape[0])
pod_id = all_pods.loc[idx, "itunes_id"]
all_pods.iloc[idx:idx+1]

results = all_matches[all_matches["itunes_id"] == pod_id][["match_id", "score"]]
results.columns = ["itunes_id", "score"]
results.merge(all_pods)[:20]

results = all_matches_meanvec[all_matches_meanvec["itunes_id"] == pod_id][["match_id", "score"]]
results.columns = ["itunes_id", "score"]
results.merge(all_pods)[:20]

all_meanvecs = pods2[["itunes_id","vector"]].merge(pods1, on = "itunes_id").merge(df, on = "itunes_id")[["itunes_id","vector"]]

all_meanvecs.to_pickle("all_mean_vectors")

all_matches_meanvec[all_matches_meanvec["itunes_id"] == pod_id]

all_pods[all_pods["itunes_id"] == 1419742734]
