==> 10_create_sql_tables.ipynb <==
import pandas as pd
from psycopg2 import connect, extensions, sql
from sqlalchemy import create_engine
import psycopg2 
import sqlalchemy
import io

def fix_desc(x):
    if x:
        return " ".join(x.split())
    else:
        return None

all_pods = pd.read_pickle("all_pods.pkl")
all_matches_episodes = pd.read_pickle("all_matches_episode_based.pkl")
all_matches_meanvec = pd.read_pickle("all_matches_mean_vec_based.pkl")

all_pods["descriptions"] = all_pods["descriptions"].apply(fix_desc)

g1 = all_matches_episodes.groupby(by="itunes_id")

all_pods.columns

matches_string = []
scores_string = []
for key, vals in g1.groups.items():
    matches_string.append(",".join([str(v) for v in all_matches_episodes.loc[vals,"match_id"].to_list()]))
    scores_string.append(",".join(["{:.4f}".format(v) for v in all_matches_episodes.loc[vals,"score"].to_list()]))

all_matches_episodes2 = pd.DataFrame(columns=["itunes_id","match_id","score"])
all_matches_episodes2["itunes_id"]=list(g1.groups.keys())
all_matches_episodes2["match_id"]=matches_string
all_matches_episodes2["score"]=scores_string
all_matches_episodes2

g2 = all_matches_meanvec.groupby(by="itunes_id")
matches_string = []
scores_string = []
for key, vals in g2.groups.items():
    matches_string.append(",".join([str(v) for v in all_matches_meanvec.loc[vals,"match_id"].to_list()]))
    scores_string.append(",".join(["{:.6f}".format(v) for v in all_matches_meanvec.loc[vals,"score"].to_list()]))

all_matches_meanvec2 = pd.DataFrame(columns=["itunes_id","match_id","score"])
all_matches_meanvec2["itunes_id"]=list(g2.groups.keys())
all_matches_meanvec2["match_id"]=matches_string
all_matches_meanvec2["score"]=scores_string
all_matches_meanvec2

all_pods

engine = create_engine('postgresql+psycopg2://naeem:mypass@localhost:5432/podcasts')

all_pods.head(0).to_sql('all_pods', engine, if_exists='replace',index=False)
conn = engine.raw_connection()
cur = conn.cursor()
output = io.StringIO()
all_pods.to_csv(output, sep='\t', header=False, index=False)
output.seek(0)
contents = output.getvalue()
cur.copy_from(output, 'all_pods', null="") # null values become ''
conn.commit()

all_matches_episodes2.head(0).to_sql('all_matches_episodes', engine, if_exists='replace',index=False)
conn = engine.raw_connection()
cur = conn.cursor()
output = io.StringIO()
all_matches_episodes2.to_csv(output, sep='\t', header=False, index=False)
output.seek(0)
contents = output.getvalue()
cur.copy_from(output, 'all_matches_episodes', null="") # null values become ''
conn.commit()

all_matches_meanvec2.head(0).to_sql('all_matches_meanvec', engine, if_exists='replace',index=False)
conn = engine.raw_connection()
cur = conn.cursor()
output = io.StringIO()
all_matches_meanvec2.to_csv(output, sep='\t', header=False, index=False)
output.seek(0)
contents = output.getvalue()
cur.copy_from(output, 'all_matches_meanvec', null="") # null values become ''
conn.commit()

## Let's see if it works...

conn = psycopg2.connect(dbname="podcasts", user="naeem", password="mypass", host="localhost", port="5432")
cursor = conn.cursor()

query = "SELECT itunes_id FROM all_pods WHERE titles in ('Fresh Air', 'This American Life');"
cursor.execute(query)
cursor.fetchall()

query = "SELECT match_id, score FROM all_matches_meanvec WHERE itunes_id = '1518397004';"
cursor.execute(query)
recom_list1 = cursor.fetchone()
recom_list1

itunes_ids = recom_list1[0].split(',')
query = "SELECT itunes_id, titles FROM all_pods WHERE itunes_id IN %s;"
cursor.execute(query, (tuple(itunes_ids),))
cursor.fetchall()[:10]
